Started breaking out parts of legend
