+++ src/pages/maps/@id/correlation/legend-item.ts
@@ -1,0 +1,5 @@
+import h from "@macrostrat/hyper";
+
+export function LegendPopup({ map }) {
+  return h("div", "Hello");
+}
